refactor(tauri): expose all items meant to be public
