+++ core/tauri-macros/src/lib.rs
@@ -26,18 +26,7 @@ pub fn generate_handler(item: TokenStream) -> TokenStream {
   gen.into()
 }
 
-/// Reads a Tauri config file and generates an [`AsTauriContext`] based on the content.
-///
-/// The default config file path is a `tauri.conf.json` file inside the Cargo manifest directory of
-/// the crate being built.
-///
-/// # Custom Config Path
-///
-/// You may pass a string literal to this macro to specify a custom path for the Tauri config file.
-/// If the path is relative, it will be search for relative to the Cargo manifest of the compiling
-/// crate.
-///
-/// todo: link the [`AsTauriContext`] docs
+/// Reads a Tauri config file and generates a `::tauri::Context` based on the content.
 #[proc_macro]
 pub fn generate_context(items: TokenStream) -> TokenStream {
   // this macro is exported from the context module
