+++ core/tauri/src/endpoints/global_shortcut.rs
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 use super::InvokeResponse;
-use crate::runtime::{window::Window, Dispatch, Params};
+use crate::{runtime::Dispatch, Params, Window};
 use once_cell::sync::Lazy;
 use serde::Deserialize;
 use std::sync::{Arc, Mutex};
