+++ core/tauri/src/hooks.rs
@@ -4,7 +4,8 @@
 
 use crate::{
   api::rpc::{format_callback, format_callback_result},
-  runtime::{app::App, window::Window, Params},
+  runtime::app::App,
+  Params, Window,
 };
 use serde::{Deserialize, Serialize};
 use std::future::Future;
