+++ core/tauri/src/lib.rs
@@ -14,7 +14,7 @@
 pub use error::Error;
 pub use tauri_api as api;
 pub(crate) use tauri_api::private::async_runtime;
-pub use tauri_macros::*;
+pub use tauri_macros::{command, generate_handler};
 
 /// The Tauri-specific settings for your runtime e.g. notification permission status.
 pub mod settings;
@@ -24,33 +24,306 @@ mod endpoints;
 mod error;
 mod event;
 mod hooks;
-/// The plugin manager module contains helpers to manage runtime plugins.
 pub mod plugin;
-/// The internal runtime between an [`App`] and the webview.
 pub mod runtime;
+#[cfg(feature = "updater")]
+mod updater;
 
-/// Tauri result type.
+/// `Result<T, ::tauri::Error>`
 pub type Result<T> = std::result::Result<T, Error>;
 
 /// A task to run on the main thread.
 pub type SyncTask = Box<dyn FnOnce() + Send>;
 
-/// types likely to be used by applications
+use crate::api::assets::Assets;
+use crate::api::config::Config;
+use crate::event::{Event, EventHandler};
+use crate::runtime::tag::Tag;
+use crate::runtime::window::PendingWindow;
+use crate::runtime::{Dispatch, Runtime};
+use serde::Serialize;
+use std::collections::HashMap;
+use std::path::PathBuf;
+
+// Export types likely to be used by the application.
 pub use {
   api::config::WindowUrl,
   hooks::InvokeMessage,
-  runtime::app::Builder,
+  runtime::app::{App, Builder},
   runtime::webview::Attributes,
-  runtime::window::Window,
-  runtime::{Context, Manager, Params},
+  runtime::window::export::Window,
 };
 
-/// Easy helper function to use the Tauri Context you made during build time.
+/// Reads the config file at compile time and generates a [`Context`] based on its content.
+///
+/// The default config file path is a `tauri.conf.json` file inside the Cargo manifest directory of
+/// the crate being built.
+///
+/// # Custom Config Path
+///
+/// You may pass a string literal to this macro to specify a custom path for the Tauri config file.
+/// If the path is relative, it will be search for relative to the Cargo manifest of the compiling
+/// crate.
+///
+/// # Note
+///
+/// This macro should not be called if you are using [`tauri-build`] to generate the context from
+/// inside your build script as it will just cause excess computations that will be discarded. Use
+/// either the [`tauri-build] method or this macro - not both.
+///
+/// [`tauri-build`]: https://docs.rs/tauri-build
+pub use tauri_macros::generate_context;
+
+/// Include a [`Context`] that was generated by [`tauri-build`] inside your build script.
+///
+/// You should either use [`tauri-build`] and this macro to include the compile time generated code,
+/// or [`generate_context!`]. Do not use both at the same time, as they generate the same code and
+/// will cause excess computations that will be discarded.
+///
+/// [`tauri-build`]: https://docs.rs/tauri-build
 #[macro_export]
 macro_rules! tauri_build_context {
   () => {
     include!(concat!(env!("OUT_DIR"), "/tauri-build-context.rs"))
   };
+}
+
+/// A icon definition.
+pub enum Icon {
+  /// Icon from file path.
+  File(PathBuf),
+  /// Icon from raw bytes.
+  Raw(Vec<u8>),
+}
+
+/// User supplied data required inside of a Tauri application.
+pub struct Context<A: Assets> {
+  /// The config the application was prepared with.
+  pub config: Config,
+
+  /// The assets to be served directly by Tauri.
+  pub assets: A,
+
+  /// The default window icon Tauri should use when creating windows.
+  pub default_window_icon: Option<Vec<u8>>,
+
+  /// Package information.
+  pub package_info: tauri_api::PackageInfo,
+}
+
+/// Types associated with the running Tauri application.
+pub trait Params: sealed::ParamsPrivate<Self> {
+  /// The event type used to create and listen to events.
+  type Event: Tag;
+
+  /// The type used to determine the name of windows.
+  type Label: Tag;
+
+  /// Assets that Tauri should serve from itself.
+  type Assets: Assets;
+
+  /// The underlying webview runtime used by the Tauri application.
+  type Runtime: Runtime;
+}
+
+/// Manages a running application.
+pub trait Manager<M: Params>: sealed::ManagerPrivate<M> {
+  /// The [`Config`] the manager was created with.
+  fn config(&self) -> &Config {
+    self.manager().config()
+  }
+
+  /// Emits a event to all windows.
+  fn emit_all<S: Serialize + Clone>(&self, event: M::Event, payload: Option<S>) -> Result<()> {
+    self.manager().emit_filter(event, payload, |_| true)
+  }
+
+  /// Emits an event to a window with the specified label.
+  fn emit_to<S: Serialize + Clone>(
+    &self,
+    label: &M::Label,
+    event: M::Event,
+    payload: Option<S>,
+  ) -> Result<()> {
+    self
+      .manager()
+      .emit_filter(event, payload, |w| w.label() == label)
+  }
+
+  /// Creates a new [`Window`] on the [`Runtime`] and attaches it to the [`Manager`].
+  fn create_window(&mut self, pending: PendingWindow<M>) -> Result<Window<M>> {
+    use sealed::RuntimeOrDispatch::*;
+
+    let labels = self.manager().labels().into_iter().collect::<Vec<_>>();
+    let pending = self.manager().prepare_window(pending, &labels)?;
+    match self.runtime() {
+      Runtime(runtime) => runtime.create_window(pending),
+      Dispatch(mut dispatcher) => dispatcher.create_window(pending),
+    }
+    .map(|window| self.manager().attach_window(window))
+  }
+
+  /// Listen to a global event.
+  fn listen_global<F>(&self, event: M::Event, handler: F) -> EventHandler
+  where
+    F: Fn(Event) + Send + 'static,
+  {
+    self.manager().listen(event, None, handler)
+  }
+
+  /// Listen to a global event only once.
+  fn once_global<F>(&self, event: M::Event, handler: F)
+  where
+    F: Fn(Event) + Send + 'static,
+  {
+    self.manager().once(event, None, handler)
+  }
+
+  /// Trigger a global event.
+  fn trigger_global(&self, event: M::Event, data: Option<String>) {
+    self.manager().trigger(event, None, data)
+  }
+
+  /// Remove an event listener.
+  fn unlisten(&self, handler_id: EventHandler) {
+    self.manager().unlisten(handler_id)
+  }
+
+  /// Fetch a single window from the manager.
+  fn get_window(&self, label: &M::Label) -> Option<Window<M>> {
+    self.manager().get_window(label)
+  }
+
+  /// Fetch all managed windows.
+  fn windows(&self) -> HashMap<M::Label, Window<M>> {
+    self.manager().windows()
+  }
+}
+
+/// Prevent implementation details from leaking out of the [`Manager`] and [`Params`] traits.
+pub(crate) mod sealed {
+  use super::Params;
+  use crate::runtime::Runtime;
+  use crate::{
+    api::{config::Config, PackageInfo},
+    event::{Event, EventHandler},
+    hooks::{InvokeMessage, PageLoadPayload},
+    runtime::window::{DetachedWindow, PendingWindow},
+    Window,
+  };
+  use serde::Serialize;
+  use std::collections::{HashMap, HashSet};
+  use uuid::Uuid;
+
+  /// private manager api
+  pub trait ParamsPrivate<M: Params>: Clone + Send + Sized + 'static {
+    /// Pass messages not handled by modules or plugins to the running application
+    fn run_invoke_handler(&self, message: InvokeMessage<M>);
+
+    /// Ran once for every window when the page is loaded.
+    fn run_on_page_load(&self, window: Window<M>, payload: PageLoadPayload);
+
+    /// Pass a message to be handled by a plugin that expects the command.
+    fn extend_api(&self, command: String, message: InvokeMessage<M>);
+
+    /// Initialize all the plugins attached to the [`Manager`].
+    fn initialize_plugins(&self) -> crate::Result<()>;
+
+    /// Prepare a [`PendingWindow`] to be created by the [`Runtime`].
+    ///
+    /// The passed labels should represent either all the windows in the manager. If the application
+    /// has not yet been started, the passed labels should represent all windows that will be
+    /// created before starting.
+    fn prepare_window(
+      &self,
+      pending: PendingWindow<M>,
+      labels: &[M::Label],
+    ) -> crate::Result<PendingWindow<M>>;
+
+    /// Attach a detached window to the manager.
+    fn attach_window(&self, window: DetachedWindow<M>) -> Window<M>;
+
+    /// Emit an event to javascript windows that pass the predicate.
+    fn emit_filter_internal<S: Serialize + Clone, F: Fn(&Window<Self>) -> bool>(
+      &self,
+      event: String,
+      payload: Option<S>,
+      filter: F,
+    ) -> crate::Result<()>;
+
+    /// Emit an event to javascript windows that pass the predicate.
+    fn emit_filter<S: Serialize + Clone, F: Fn(&Window<M>) -> bool>(
+      &self,
+      event: M::Event,
+      payload: Option<S>,
+      predicate: F,
+    ) -> crate::Result<()>;
+
+    /// All current window labels existing.
+    fn labels(&self) -> HashSet<M::Label>;
+
+    /// The configuration the [`Manager`] was built with.
+    fn config(&self) -> &Config;
+
+    /// App package information.
+    fn package_info(&self) -> &PackageInfo;
+
+    /// Remove the specified event handler.
+    fn unlisten(&self, handler_id: EventHandler);
+
+    /// Trigger an event.
+    fn trigger(&self, event: M::Event, window: Option<M::Label>, data: Option<String>);
+
+    /// Set up a listener to an event.
+    fn listen<F: Fn(Event) + Send + 'static>(
+      &self,
+      event: M::Event,
+      window: Option<M::Label>,
+      handler: F,
+    ) -> EventHandler;
+
+    /// Set up a listener to and event that is automatically removed after called once.
+    fn once<F: Fn(Event) + Send + 'static>(
+      &self,
+      event: M::Event,
+      window: Option<M::Label>,
+      handler: F,
+    );
+
+    fn event_listeners_object_name(&self) -> String;
+    fn event_queue_object_name(&self) -> String;
+    fn event_emit_function_name(&self) -> String;
+
+    /// Generate a random salt and store it in the manager
+    fn generate_salt(&self) -> Uuid;
+
+    /// Verify that the passed salt is a valid salt in the manager.
+    fn verify_salt(&self, salt: String) -> bool;
+
+    /// Get a single managed window.
+    fn get_window(&self, label: &M::Label) -> Option<Window<M>>;
+
+    /// Get all managed windows.
+    fn windows(&self) -> HashMap<M::Label, Window<M>>;
+  }
+
+  /// Represents either a running [`Runtime`] or a dispatcher to it.
+  pub enum RuntimeOrDispatch<'m, M: Params> {
+    /// Mutable reference to the running [`Runtime`].
+    Runtime(&'m mut M::Runtime),
+
+    /// A dispatcher to the running [`Runtime`].
+    Dispatch(<M::Runtime as Runtime>::Dispatcher),
+  }
+
+  /// Represents a managed handle to the application runner.
+  pub trait ManagerPrivate<M: Params> {
+    /// The manager behind the [`Managed`] item.
+    fn manager(&self) -> &M;
+
+    /// The runtime or runtime dispatcher of the [`Managed`] item.
+    fn runtime(&mut self) -> RuntimeOrDispatch<'_, M>;
+  }
 }
 
 #[cfg(test)]
